Add and include neco library to mt-redis
Modify Makefile in /src and /deps to compile and link neco to mt-redis.

There can be no more than one scheduler per thread, when the first
coroutine is started using neco_start(), a new Neco runtime is
initialized in the current thread, and each runtime has its own
scheduler. I put the event loop in a coroutine, when a thread start to
run the event loop, it will initialize the coroutine scheduler. When
there is a new file event, I start a new coroutine no matter what kind
of file event it is.